+++ lib/geneva_drive.rb
@@ -73,7 +73,7 @@ module GenevaDrive
   end
 
   # Set default configuration values
-  self.delete_completed_workflows_after = nil  # Disabled by default
+  self.delete_completed_workflows_after = 30.days
   self.stuck_in_progress_threshold = 1.hour
   self.stuck_scheduled_threshold = 1.hour
   self.housekeeping_batch_size = 1000